This button does something
I created a button that does something.  It was supposed to be a waving "hello" emoji, but the gif is acting like a still image.  I used mostly a walkthrough from this page to get it set up: https://developer.android.com/courses/android-basics-compose/unit-1?authuser=1

Then I used gpt to help me make a button under @Composable.  The Android Studio IDE did a lot of the work too. I still need to go back and write comments and run it with codechateditor, but I was learning as I went.